Change msg type to Uint32

+++ rb_controller/robot_controller.py
@@ -2,7 +2,7 @@ import rclpy
 import time
 from rclpy.node import Node
 
-from std_msgs.msg import String, Bool, UInt8
+from std_msgs.msg import String, Bool, UInt32
 from sensor_msgs.msg import Joy
 
 
@@ -32,7 +32,7 @@ class RobotController(Node):
         # Create publisher
         self.idle_bool_pub_ = self.create_publisher(Bool, 'idle_bool', 10)
         self.robot_state_pub_ = self.create_publisher(String, 'robot_mode', 10)
-        self.led_state_pub_ = self.create_publisher(UInt8, 'led_state', 10)
+        self.led_state_pub_ = self.create_publisher(UInt32, 'led_state', 10)
         # Create subscriber
         self.joy_sub_ = self.create_subscription(Joy, 'joy', self.button_read, 10)
         self.joy_sub_  # prevent unused variable warning
@@ -92,10 +92,11 @@ class RobotController(Node):
                 self.led_state = 2
         else:
             idle_msg.data = False
-            # 
+            # Set Start Time (To calculate target lost)
             if self.robot_state == 'FOLLOW' and self.target_status == True:
                 self.timer0 = round(time.time(), 0)
                 self.timer1 = self.timer0 + 5.0
+            # Target lost Time Count
             elif self.robot_state == 'FOLLOW' and self.target_status == False:
                 if round(time.time(), 0) == self.timer1:
                     self.robot_state = 'IDLE'
@@ -109,7 +110,7 @@ class RobotController(Node):
         mode_msg.data = self.robot_state
         self.robot_state_pub_.publish(mode_msg)
         # Publish LED states
-        led_msg = UInt8()
+        led_msg = UInt32()
         led_msg.data = self.led_state
         self.led_state_pub_.publish(led_msg)
         # Publish Idle message
